requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
pymongo==4.5.0
motor==3.3.2
python-dateutil==2.8.2
dnspython==2.6.1
